+++ src/main/resources/release_notes.md
@@ -35,6 +35,7 @@
 ### Youth
 * fix bug in potential calculation due to wrong isTop3 skill tagging (#1278)
 * fix escaping of youth team names (#1295) 
+* fix set pieces is not limited by top3 skill rules (#1309)
 
 ### Misc